@@ -5,13 +5,13 @@ version: 0.0.2
 homepage: https://github.com/flymzero/webdav_client
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  dio: ^3.0.10
-  xml: ^4.5.1
-  meta: ^1.2.3
-  convert: ^2.1.1
+  dio: ^4.0.0-beta6
+  xml: ^5.0.2
+  meta: ^1.3.0
+  convert: ^3.0.0
 
 dev_dependencies:
   test: '>=1.0.0 <2.0.0'
